Dungeon NextRoom/Exit
Added a restriction on creating dungeons : need a room of type exit to be able to create the dungeon. The player will exit on this room, so further rooms are useless.
Enter command to enter the dungeon.
EnterRoom to go to the next room if cleared.
Exit now properly exits. If not on exit type room, you will not get any exp.
Closes #41, closes #46